+++ client/src/components/FiltersSectionComponent.js
