+++ client/src/components/ProductsGridComponent.js
