+++ client/src/components/pages/ProductDetail.js
@@ -4,7 +4,7 @@ import { useParams } from 'react-router-dom';
 import { Link } from 'react-router-dom';
 import Carousel from 'react-material-ui-carousel';
 
-import Navbar from '../Navbar.js';
+import Navbar from '../NavbarComponent.js';
 import Products from '../AllProducts.js';
 
 import ArrowNext from "../../assets/next-arrow.svg";
